project3/requirements.txt
numpy
ipywidgets==7.6.5
ipython
Pillow